+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.21.0] - 2025-12-14
+
+### Added
+
+- `Parser` and `AsyncParser` to represent parsing functions.
+- `Transform` and `AsyncTransform` to represent transformations.
+
 ## [0.20.0] - 2025-12-06
 
 ### Added
@@ -233,6 +240,8 @@ that does the same, but probably better.
   - `isNumber`
   - `isFunction`
 
+[0.21.0]: https://github.com/infra-blocks/ts-types/compare/v0.20.0...v0.21.0
+[0.20.0]: https://github.com/infra-blocks/ts-types/compare/v0.19.0...v0.20.0
 [0.19.0]: https://github.com/infra-blocks/ts-types/compare/v0.18.0...v0.19.0
 [0.18.0]: https://github.com/infra-blocks/ts-types/compare/v0.17.0...v0.18.0
 [0.17.0]: https://github.com/infra-blocks/ts-types/compare/v0.16.0...v0.17.0